--- docs/src/jekyll/.gitignore
@@ -1,1 +1,0 @@
-_site
