--- docs/src/jekyll/Gemfile
@@ -1,6 +1,0 @@
-source 'https://rubygems.org'
-
-gem "jekyll"
-gem "redcarpet"
-gem "rouge"
-gem "compass"
