+++ docs/src/main/tut/index.md
@@ -3,6 +3,21 @@ layout: default
 title: http4s
 ---
 
+```tut:silent
+import org.http4s._
+import org.http4s.dsl._
+import org.http4s.server._
+import org.http4s.server.websocket._
+import org.http4s.websocket.WebsocketBits._
+import org.http4s.server.blaze._
+
+import scalaz.concurrent._
+import scala.concurrent.duration._
+import scalaz.stream._
+import scalaz.stream.async.unboundedQueue
+import scalaz.stream.time.awakeEvery
+```
+
 http4s is a minimal, idiomatic Scala interface for HTTP.  http4s is Scala's answer to Ruby's 
 Rack, Python's WSGI, Haskell's WAI, and Java's Servlets.
 
@@ -26,7 +41,11 @@ Building on the FP tools of scalaz not only makes an `HttpService` simple to def
 it also makes them easy to compose.  Adding gzip compression or rewriting URIs is
 as simple as applying a middleware to an `HttpService`.
 
-{%code_ref ../test/scala/org/http4s/docs/CompositionExample.scala composition_example %}
+```tut:silent
+val service = HttpService { case req => Ok("Foo") }
+val wcompression = middleware.GZip(service)
+val translated   = middleware.URITranslation.translateRoot("/http4s")(service)
+```
 
 ### Asynchronous
 
@@ -34,11 +53,49 @@ Any http4s response can be streamed from an asynchronous source. http4s offers a
 of helpers to help you get your data out the door in the fastest way possible without
 tying up too many threads.
 
-{%code_ref ../test/scala/org/http4s/docs/AsynchronousExample.scala asynchronous_example %}
+```tut:silent
+// Make your model safe and streaming by using a scalaz-stream Process
+def getData(req: Request): Process[Task, String] = Process("one", "two", "three")
+
+val service = HttpService.apply {
+  // Wire your data into your service
+  case req@GET -> Root / "streaming" => Ok(getData(req))
+  // You can use helpers to send any type of data with an available EntityEncoder[T]
+  case GET -> Root / "synchronous" => Ok("This is good to go right now.")
+}
+```
 
 http4s is a forward-looking technology.  HTTP/2.0 and WebSockets will play a central role.
 
-{%code_ref ../../../examples/blaze/src/main/scala/com/example/http4s/blaze/BlazeWebSocketExample.scala blaze_websocket_example %}
+```tut:silent
+val route = HttpService {
+  case GET -> Root / "hello" =>
+    Ok("Hello world.")
+
+  case req@ GET -> Root / "ws" =>
+    val src = awakeEvery(1.seconds)(Strategy.DefaultStrategy, DefaultScheduler).map{ d => Text(s"Ping! $d") }
+    val sink: Sink[Task, WebSocketFrame] = Process.constant {
+      case Text(t, _) => Task.delay( println(t))
+      case f       => Task.delay(println(s"Unknown type: $f"))
+    }
+    WS(Exchange(src, sink))
+
+  case req@ GET -> Root / "wsecho" =>
+    val q = unboundedQueue[WebSocketFrame]
+    val src = q.dequeue.collect {
+      case Text(msg, _) => Text("You sent the server: " + msg)
+    }
+
+    WS(Exchange(src, q.enqueue))
+}
+
+```scala
+BlazeBuilder.bindHttp(8080)
+  .withWebSockets(true)
+  .mountService(route, "/http4s")
+  .run
+  .awaitShutdown()
+```
 
 ## Choose your backend
 
@@ -48,7 +105,14 @@ needs now, and easily port if and when your needs change.
 
 [blaze](http://github.com/http4s/blaze) is an NIO framework.  Run http4s on blaze for maximum throughput.
 
-{%code_ref ../../../examples/blaze/src/main/scala/com/example/http4s/blaze/BlazeExample.scala blaze_server_example %}
+```scala
+object BlazeExample extends App {
+  BlazeBuilder.bindHttp(8080)
+    .mountService(ExampleService.service, "/http4s")
+    .run
+    .awaitShutdown()
+}
+```
 
 ### Servlets
 
@@ -57,18 +121,67 @@ on your existing infrastructure, and take full advantage of the mature JVM ecosy
 http4s can run in a .war on any Servlet 3.0+ container, and comes with convenient builders
 for embedded Tomcat and Jetty containers.
 
-{%code_ref ../../../examples/jetty/src/main/scala/com/example/http4s/jetty/JettyExample.scala jetty_example %}
+```scala
+object JettyExample extends App {
+  val metrics = new MetricRegistry
+
+  JettyBuilder
+    .bindHttp(8080)
+    .withMetricRegistry(metrics)
+    .mountService(ExampleService.service, "/http4s")
+    .mountServlet(new MetricsServlet(metrics), "/metrics/*")
+    .run
+    .awaitShutdown()
+}
+```
 
 ## An Asynchronous Client ##
 
 http4s also offers an asynchronous HTTP client built on the same model as the server.
 
-{%code_ref ../../../examples/blaze/src/main/scala/com/example/http4s/blaze/ClientExample.scala blaze_client_example %}
+```tut:silent
+import org.http4s.Http4s._
+import org.http4s.client.blaze._
+import scalaz.concurrent.Task
+
+val client = PooledHttp1Client()
+
+val page: Task[String] = client.getAs[String](uri("https://www.google.com/"))
+
+for (_ <- 1 to 2)
+  println(page.run.take(72))   // each execution of the Task will refetch the page!
+
+// We can do much more: how about decoding some JSON to a scala object
+// after matching based on the response status code?
+import org.http4s.Status.NotFound
+import org.http4s.Status.ResponseClass.Successful
+import _root_.argonaut.DecodeJson
+import org.http4s.argonaut.jsonOf
+
+case class Foo(bar: String)
+
+implicit val fooDecode = DecodeJson(c => for { // Argonaut decoder. Could also use json4s.
+  bar <- (c --\ "bar").as[String]
+} yield Foo(bar))
+
+// jsonOf is defined for Json4s and Argonaut, just need the right decoder!
+implicit val fooDecoder = jsonOf[Foo]
+
+// Match on response code!
+val page2 = client.get(uri("http://http4s.org/resources/foo.json")) {
+  case Successful(resp) => resp.as[Foo].map("Received response: " + _)
+  case NotFound(resp)   => Task.now("Not Found!!!")
+  case resp             => Task.now("Failed: " + resp.status)
+}
+
+println(page2.run)
+
+client.shutdown.run
+```
 
 ## Other features ##
 
 * [twirl](https://github.com/playframework/twirl) integration: use Play framework templates with http4s
-
 
 ## Projects using http4s ##
