+++ docs/src/site/_config.yml
@@ -1,0 +1,5 @@
+name: http4s.org
+markdown: redcarpet
+highlighter: pygments
+baseurl: /http4s
+apidocs: /http4s/api
