+++ docs/src/site/resources/foo.json
