+++ project/plugins.sbt
@@ -10,5 +10,6 @@ addSbtPlugin("com.typesafe.sbt" % "sbt-site" % "0.8.2")
 addSbtPlugin("com.typesafe.sbt" % "sbt-twirl" % "1.1.1")
 addSbtPlugin("io.gatling" % "gatling-sbt" % "2.1.5")
 addSbtPlugin("io.spray" % "sbt-revolver" % "0.8.0")
+addSbtPlugin("org.tpolecat" % "tut-plugin" % "0.4.2")
 addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.2.6")
 addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.0.0")
